RavenDB-5977 - More understandable read value when compressing in batches - Detect memory corruption more easily - Avoid int overflow resulting in negative values when compressing very large values - Take into account that the TransactionHeader also require a size that may be a page - Ensure that we don't have int overflow when writing to the pager -

+++ src/Voron/Impl/Paging/AbstractPager.cs
@@ -298,7 +298,7 @@ namespace Voron.Impl.Paging
                 numberOfPages = this.GetNumberOfOverflowPages(pageHeader->OverflowSize);
             }
             const int adjustPageSize = (Constants.Storage.PageSize)/(4*Constants.Size.Kilobyte);
-            destwI4KbBatchWrites.Write(pageHeader->PageNumber * adjustPageSize, numberOfPages * adjustPageSize, src);
+            destwI4KbBatchWrites.Write(pageHeader->PageNumber * (long)adjustPageSize, numberOfPages * adjustPageSize, src);
 
             return numberOfPages;
         }
